pyproject.toml
[project]
dynamic = [
    "version"
]
name = "seis-lab-data"
description = "marine data catalog"
readme = "README.md"
authors = [
    { name = "Ricardo Garcia Silva", email = "ricardo.garcia@naturalgis.pt" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "authlib>=1.6.1",
    "dramatiq[redis,watch]>=1.18.0",
    "geoalchemy2>=0.18.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "redis[hiredis]>=6.2.0",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
    "starlette>=0.47.1",
    "starlette-babel>=1.0.3",
    "typer>=0.16.0",
    "uvicorn[standard]>=0.35.0",
]
[project.scripts]
seis-lab-data = "seis_lab_data.main:app"

[dependency-groups]
dev = [
    "playwright>=1.53.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-playwright>=0.7.0",
    "ruff>=0.12.1",
]
jupyter = [
    "jupyterlab>=4.4.4",
]

[tool.uv]
default-groups = []
package = true

[tool.pytest.ini_options]
addopts = "--strict-markers -m 'not e2e'"
markers = [
    "e2e: marks tests as being end-to-end tests",
]
